photogallery/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorthief==0.2.1
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.2.0
psycopg2==2.9.5
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
